+++ e3nn/o3/_norm.py
@@ -23,6 +23,9 @@ class Norm(CodeGenMixin, torch.nn.Module):
     epsilon : float, optional
         if not ``None``, norms smaller than ``epsilon`` will be "rounded" up to ``epsilon`` *before* any square roots are taken. (Doing the masking internally in ``Norm`` this way prevents NaNs in gradients.)
 
+    squared : bool, optional
+        Whether to return the squared norm. ``False`` by default, i.e. the norm itself (sqrt of squared norm) is returned. Setting this to ``True`` in situations where the extra power of two doesn't matter can help avoid NaNs without ``epsilon``.
+
     Examples
     --------
     Compute the norms of 17 vectors.
@@ -32,8 +35,9 @@ class Norm(CodeGenMixin, torch.nn.Module):
     torch.Size([17])
     """
     epsilon: Optional[float]
+    squared: bool
 
-    def __init__(self, irreps_in, epsilon: Optional[float] = None):
+    def __init__(self, irreps_in, epsilon: Optional[float] = None, squared: bool = False):
         super().__init__()
         self.irreps_in = o3.Irreps(irreps_in)
         self.irreps_out = o3.Irreps([(mul, "0e") for mul, _ in self.irreps_in]).simplify()
@@ -42,8 +46,11 @@ class Norm(CodeGenMixin, torch.nn.Module):
             epsilon = float(epsilon)
             assert epsilon > 0
         self.epsilon = epsilon
+        self.squared = squared
+        if squared and epsilon is not None:
+            raise ValueError("epsilon doesn't make sense if `squared` is True")
 
-        code, indptr = _codegen_norm(self.irreps_in, self.epsilon)
+        code, indptr = _codegen_norm(self.irreps_in, self.epsilon, self.squared)
         self._codegen_register({
             "_compiled_main": code
         })
@@ -68,7 +75,7 @@ class Norm(CodeGenMixin, torch.nn.Module):
         return self._compiled_main(features, self._indptr)
 
 
-def _codegen_norm(irreps_in: o3.Irreps, epsilon: Optional[float]) -> Tuple[str, torch.Tensor]:
+def _codegen_norm(irreps_in: o3.Irreps, epsilon: Optional[float], squared: bool) -> Tuple[str, torch.Tensor]:
     graph_out = fx.Graph()
 
     # = Function definitions =
@@ -121,7 +128,11 @@ def _codegen_norm(irreps_in: o3.Irreps, epsilon: Optional[float]) -> Tuple[str, 
         for_sqrt.masked_fill_(out < epsilon, epsilon)
     else:
         for_sqrt = out
-    out = torch.sqrt(for_sqrt)
+
+    if not squared:
+        out = torch.sqrt(for_sqrt)
+    else:
+        out = for_sqrt
 
     # = Return the result =
     out = out.reshape(outsize)