+++ tests/o3/norm_test.py
@@ -8,7 +8,7 @@ from e3nn.util.test import assert_equivariant, assert_auto_jitable, random_irrep
 
 class SlowNorm(torch.nn.Module):
     r"""Slow norm using TensorProduct"""
-    def __init__(self, irreps_in, epsilon=None):
+    def __init__(self, irreps_in, epsilon=None, squared=False):
         super().__init__()
 
         irreps_in = o3.Irreps(irreps_in).simplify()
@@ -30,6 +30,7 @@ class SlowNorm(torch.nn.Module):
         self.irreps_in = irreps_in
         self.irreps_out = irreps_out.simplify()
         self.epsilon = None if epsilon is None else float(epsilon)
+        self.squared = squared
 
     def forward(self, features):
         out = self.tp(features, features)
@@ -39,18 +40,21 @@ class SlowNorm(torch.nn.Module):
             for_sqrt[out < eps_squared] = eps_squared
         else:
             for_sqrt = out
-        return for_sqrt.sqrt()
+        if self.squared:
+            return for_sqrt
+        else:
+            return for_sqrt.sqrt()
 
 
 @pytest.mark.parametrize(
     "irreps_in", ["", "5x0e", "1e + 2e + 4x1e + 3x3o"] + random_irreps(n=4)
 )
 @pytest.mark.parametrize("batchdim", [(4,), (1,), tuple(), (5, 3, 7)])
-@pytest.mark.parametrize("eps", [None, 1e-3])
-def test_norm_like_tp(irreps_in, batchdim, eps):
+@pytest.mark.parametrize("eps, squared", [(None, True), (None, False), (1e-3, False)])
+def test_norm_like_tp(irreps_in, batchdim, eps, squared):
     """Test that Norm gives the same results as the corresponding TensorProduct."""
-    m = o3.Norm(irreps_in, epsilon=eps)
-    m_true = SlowNorm(irreps_in, epsilon=eps)
+    m = o3.Norm(irreps_in, epsilon=eps, squared=squared)
+    m_true = SlowNorm(irreps_in, epsilon=eps, squared=squared)
     inp = torch.randn(batchdim + (m.irreps_in.dim,))
     out = m(inp)
     out_true = m_true(inp)